Correctly setup roles graph for children roles
First $children prop. in Role class must be initialized to
an empty array.
Second provided children roles must be an array. Then this two
arrays get merged.

Last, when role is added it is checked for presence of children.
If role has children, those are removed from current roles graph
and readded again with parent set to current role.